backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==2.3.0

# Database
psycopg2-binary==2.9.10
redis==5.0.1

# Environment & Security
python-decouple==3.8
argon2-cffi==23.1.0
cryptography==41.0.7

# Image Processing
Pillow>=10.2.0
boto3==1.29.7  # For AWS S3

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Development Tools
django-extensions==3.2.3
django-debug-toolbar==4.2.0
ipython==8.17.2

# API Documentation
drf-spectacular==0.26.5

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Utilities
django-storages==1.14.2
django-filter==23.4
